+++ Data Scraping/data/ingredient_links.txt
@@ -1,29 +1,2 @@
 https://www.food.com/about/dill-164
 https://www.food.com/about/parsley-171
-https://www.food.com/about/green-onion-363
-https://www.food.com/about/mint-330
-https://www.food.com/about/salt-359
-https://www.food.com/about/lemon-125
-https://www.food.com/about/mint-330
-https://www.food.com/about/dill-164
-https://www.food.com/about/garlic-165
-https://www.food.com/about/cider-vinegar-525
-https://www.food.com/about/salt-359
-https://www.food.com/about/pepper-337
-https://www.food.com/about/olive-oil-495
-https://www.food.com/about/paprika-335
-https://www.food.com/about/tomato-151
-https://www.food.com/about/olive-oil-495
-https://www.food.com/about/salt-359
-https://www.food.com/about/eggplant-128
-https://www.food.com/about/onion-148
-https://www.food.com/about/garlic-165
-https://www.food.com/about/paprika-335
-https://www.food.com/about/cayenne-pepper-320
-https://www.food.com/about/cumin-20
-https://www.food.com/about/sweet-potato-307
-https://www.food.com/about/zucchini-316
-https://www.food.com/about/sweet-pepper-150
-https://www.food.com/about/tortilla-657
-https://www.food.com/about/egg-142
-https://www.food.com/about/avocado-190
